requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
PyJWT==2.8.0
cryptography==41.0.7
